pyproject.toml
[tool.poetry]
name = "discordpy-twitter"
version = "0.1.0"
description = ""
authors = ["rokkish <rokki.hobby@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
mypy = "^1.6.1"
black = "^23.10.1"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
"discord.py" = "^2.3.2"
isort = "^5.12.0"
python-dotenv = "^1.0.0"
tweepy = "^4.14.0"
bs4 = "^0.0.1"
feedparser = "^6.0.10"
sqlalchemy = "^2.0.22"
alembic = "^1.12.0"
pytest-mock = "^3.12.0"
mecab = "^0.996.3"
wikipedia = "^1.4.0"
pydantic = "^2.6.2"
python-levenshtein = "^0.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
